WIP: Refactor code. (Not Building)
