+++ krystal-common/src/main/java/com/flipkart/krystal/pooling/Lease.java
@@ -1,4 +1,4 @@
-package com.flipkart.krystal.utils;
+package com.flipkart.krystal.pooling;
 
 import org.checkerframework.checker.nullness.qual.NonNull;
