+++ krystal-common/src/main/java/com/flipkart/krystal/pooling/MultiLeasePolicy.java
@@ -1,3 +1,3 @@
-package com.flipkart.krystal.utils;
+package com.flipkart.krystal.pooling;
 
 public sealed interface MultiLeasePolicy permits PreferObjectReuse, DistributeLeases {}
