+++ krystal-common/src/main/java/com/flipkart/krystal/pooling/PreferObjectReuse.java
@@ -1,4 +1,4 @@
-package com.flipkart.krystal.utils;
+package com.flipkart.krystal.pooling;
 
 import java.util.Optional;
