+++ vajram/vajram-samples/src/test/java/com/flipkart/krystal/vajram/samples/calculator/Add2And3Test.java
@@ -5,8 +5,8 @@ import static com.flipkart.krystal.vajram.VajramID.ofVajram;
 import com.flipkart.krystal.executors.SingleThreadExecutor;
 import com.flipkart.krystal.executors.ThreadPerRequestPool;
 import com.flipkart.krystal.krystex.kryon.KryonExecutorConfig;
-import com.flipkart.krystal.utils.Lease;
-import com.flipkart.krystal.utils.LeaseUnavailableException;
+import com.flipkart.krystal.pooling.Lease;
+import com.flipkart.krystal.pooling.LeaseUnavailableException;
 import com.flipkart.krystal.vajramexecutor.krystex.KrystexVajramExecutor;
 import com.flipkart.krystal.vajramexecutor.krystex.KrystexVajramExecutorConfig;
 import com.flipkart.krystal.vajramexecutor.krystex.VajramKryonGraph;
